fix : 404 찍고 렌더링 되는 부분 수정
